Pmnps@4.3.5

+++ platforms/pmnps/src/libs/file/index.ts
@@ -169,6 +169,9 @@ async function rename(
 }
 
 async function rmdir(filePath: string): Promise<boolean> {
+  if(!fs.existsSync(filePath)){
+    return false;
+  }
   return new Promise((resolve, reject) => {
     fs.rm(filePath, { recursive: true }, err => {
       if (err) {
@@ -305,7 +308,13 @@ async function copyFolder(
     fs.cp(
       sourceDirPath,
       targetDirPath,
-      { recursive: true, force: false },
+      {
+        recursive: true,
+        force: false,
+        filter(source: string): boolean | Promise<boolean> {
+          return source !== 'node_modules';
+        }
+      },
       err => {
         if (err) {
           reject(err);
